# src/main/java/com/lms/real_estate_crm_controller/LeadsController.java
// java
package com.lms.real_estate_crm_controller;

import com.lms.crm.api.leads.LeadsApi;
import com.lms.crm.model.leads.Lead;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.RestController;

import jakarta.validation.Valid;
import org.springframework.format.annotation.DateTimeFormat;

import java.time.LocalDate;
import java.util.*;
import java.util.concurrent.ConcurrentHashMap;

/*
 Brief behavior:
 - Uses an in-memory ConcurrentHashMap to store Lead objects by id.
 - createLead: assigns a UUID if id is missing, stores lead, returns 201.
 - deleteLead: removes lead if exists, returns 204 or 404.
 - getLeadById: returns 200 if found, 404 otherwise (bodyless because the generated interface uses ResponseEntity<Void>).
 - getLeads: returns 200 (bodyless) and supports paging params but does not return actual list due to interface signature.
 - leadComplianceReport: returns 200 (bodyless).
 - updateLead: updates stored lead if exists, returns 200 or 404.
 Note: The generated `LeadsApi` methods return ResponseEntity<Void>. Adjust signatures or generated API if you need bodies returned.
*/

@RestController
@Validated
public class LeadsController implements LeadsApi {

    private final Map<String, Lead> store = new ConcurrentHashMap<>();

    @Override
    public Optional<org.springframework.web.context.request.NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    @Override
    public ResponseEntity<Void> createLead(@Valid @org.springframework.web.bind.annotation.RequestBody Lead lead) {
        if (lead == null) {
            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
        }

        // assign id if missing (assumes Lead has setId/getId)
        try {
            String id = null;
            try {
                id = (String) Lead.class.getMethod("getId").invoke(lead);
            } catch (NoSuchMethodException ignored) { /* model may use different naming; best-effort */ }

            if (id == null || id.isBlank()) {
                String generated = UUID.randomUUID().toString();
                try {
                    Lead.class.getMethod("setId", String.class).invoke(lead, generated);
                } catch (NoSuchMethodException ignored) { /* ignore if setter absent */ }
            }
        } catch (Exception ignored) {
            // if reflection fails, continue; assume caller provided id
        }

        // attempt to obtain id to store
        String storeId = null;
        try {
            Object idObj = Lead.class.getMethod("getId").invoke(lead);
            if (idObj != null) storeId = idObj.toString();
        } catch (Exception ignored) {}

        if (storeId == null || storeId.isBlank()) {
            storeId = UUID.randomUUID().toString();
            try {
                Lead.class.getMethod("setId", String.class).invoke(lead, storeId);
            } catch (Exception ignored) {}
        }

        store.put(storeId, lead);
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

    @Override
    public ResponseEntity<Void> deleteLead(@org.springframework.web.bind.annotation.PathVariable("id") String id) {
        if (id == null || !store.containsKey(id)) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
        }
        store.remove(id);
        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
    }

    @Override
    public ResponseEntity<Void> getLeadById(@org.springframework.web.bind.annotation.PathVariable("id") String id) {
        if (id == null || !store.containsKey(id)) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
        }
        // Interface expects ResponseEntity<Void>; returning status 200 without body.
        return ResponseEntity.ok().build();
    }

    @Override
    public ResponseEntity<Void> getLeads(
            @org.springframework.web.bind.annotation.RequestParam(value = "projectId", required = false) String projectId,
            @org.springframework.web.bind.annotation.RequestParam(value = "leadQuality", required = false) String leadQuality,
            @org.springframework.web.bind.annotation.RequestParam(value = "status", required = false) String status,
            @org.springframework.web.bind.annotation.RequestParam(value = "fromDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate fromDate,
            @org.springframework.web.bind.annotation.RequestParam(value = "toDate", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate toDate,
            @org.springframework.web.bind.annotation.RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
            @org.springframework.web.bind.annotation.RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
        // This implementation ignores filters because the generated interface returns Void.
        // You can implement filtering and return actual payload by updating the generated API types.
        return ResponseEntity.ok().build();
    }

    @Override
    public ResponseEntity<Void> leadComplianceReport(
            @org.springframework.web.bind.annotation.RequestParam(value = "projectId", required = false) String projectId,
            @org.springframework.web.bind.annotation.RequestParam(value = "salesUserId", required = false) String salesUserId
    ) {
        // placeholder for compliance logic
        return ResponseEntity.ok().build();
    }

    @Override
    public ResponseEntity<Void> updateLead(
            @org.springframework.web.bind.annotation.PathVariable("id") String id,
            @Valid @org.springframework.web.bind.annotation.RequestBody Lead lead
    ) {
        if (id == null || !store.containsKey(id)) {
            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
        }
        // ensure id consistency
        try {
            Lead.class.getMethod("setId", String.class).invoke(lead, id);
        } catch (Exception ignored) {}

        store.put(id, lead);
        return ResponseEntity.ok().build();
    }
}