# src/main/resources/openapi/leads.yaml
openapi: 3.0.3
info:
  title: Leads Module
  version: 1.0.0

servers:
  - url: /crm/api/v1

tags:
  - name: Leads
    description: Lead management APIs

security:
  - bearerAuth: []

paths:
  /leads:
    post:
      summary: Create Lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lead"
      responses:
        "201":
          description: Lead created
    get:
      summary: Get Leads with Filters
      operationId: getLeads
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: leadQuality
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lead list

  /leads/{id}:
    get:
      summary: Get Lead by ID
      operationId: getLeadById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead details
    put:
      summary: Update Lead
      operationId: updateLead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Lead"
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete Lead
      operationId: deleteLead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  /leads/compliance:
    get:
      summary: Lead Process Compliance Report
      operationId: leadComplianceReport
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
        - name: salesUserId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Compliance data

components:
  schemas:
    Lead:
      type: object
      required:
        - leadName
        - source
        - contactNo
      properties:
        id:
          type: string
        leadId:
          type: string
          example: LD001
        source:
          type: string
          example: Instagram
        leadName:
          type: string
        email:
          type: string
          format: email
        contactNo:
          type: string
        whatsappNo:
          type: string
        configuration:
          type: string
          example: 2 BHK
        leadQuality:
          type: string
          example: Qualified
        status:
          type: string
          example: Positive
        projectId:
          type: string
        projectOwnerId:
          type: string
        lastContactDate:
          type: string
          format: date
        nextContactDate:
          type: string
          format: date
        comments:
          type: string
        crossPitchProjects:
          type: array
          items:
            type: string
        campaignName:
          type: string
        landingPageUrl:
          type: string